update the chatbot api to use the one key
