+++ intelliserver/api/functions/chatbot/index.js
@@ -5,47 +5,7 @@ const router = express.Router();
 const { ChatGPTInput, LLamaReplicateInput, LLamaSageInput, CohereInput } = require('intellinode');
 const { Chatbot, SupportedChatModels } = require('intellinode');
 const { USE_DEFAULT_KEYS } = require(path.join(global.__basedir, 'config'));
-
-const keys = {
-    'openai': process.env.OPENAI_API_KEY,
-    'replicate': process.env.REPLICATE_API_KEY,
-    'sagemaker': process.env.SAGEMAKER_API_KEY,
-    'cohere': process.env.COHERE_API_KEY
-}
-
-function getChatbot(req) {
-    console.log('getChatbot provider: ', req.body.provider)
-    if (USE_DEFAULT_KEYS && !req.body.api_key) {
-        return new Chatbot(keys[req.body.provider.toLowerCase()], req.body.provider);
-    } else {
-        return new Chatbot(req.body.api_key, req.body.provider);
-    }
-}
-
-function addMessages(chatInput, messages) {
-    messages.forEach(msg => {
-        if(msg.role.toLowerCase() === "user") {
-            chatInput.addUserMessage(msg.content);
-        }
-        else if (msg.role.toLowerCase() === "assistant") {
-            chatInput.addAssistantMessage(msg.content);
-        }
-        else {
-            throw new Error(`Invalid message role: ${msg.role}`);
-        }
-    });
-    return chatInput;
-}
-
-function getChatInput(input, provider) {
-    const inputInst = input instanceof ChatGPTInput ? input :
-        provider === SupportedChatModels.REPLICATE ? new LLamaReplicateInput(input.system, input) :
-        provider === SupportedChatModels.SAGEMAKER ? new LLamaSageInput(input.system, input) :
-        provider === SupportedChatModels.COHERE ? new CohereInput(input.system, input) :
-        new ChatGPTInput(input.system, input);
-
-    return addMessages(inputInst, input.messages);
-}
+const ChatbotHelpers = require('../../utils/chatbot_helper');
 
 /**
  * @swagger
@@ -104,14 +64,13 @@ function getChatInput(input, provider) {
  */
 
 router.post('/chat', async (req, res, next) => {
-    console.log("this is result",req)
     try {
-        const chatbot = getChatbot(req);
-        const input = getChatInput(req.body.input, req.body.provider);
+        const chatbot = ChatbotHelpers.getChatbot(req);
+        const input = ChatbotHelpers.getChatInput(req.body.input, req.body.provider);
         const functions = req.body.functions;
         const function_call = req.body.function_call;
         const results = await chatbot.chat(input, functions, function_call);
-        res.json({ status: "OK", data: results });
+        res.json({ status: "OK", data: results });  
     } catch (error) {
         console.log("this is error",error);
         res.json({ status: "ERROR", message: error.message });
@@ -126,10 +85,9 @@ router.post('/stream', async (req, res, next) => {
         if (SupportedChatModels.OPENAI != provider.toLowerCase()) {
             res.json({ status: "ERROR", message: "stream support openai only, for other providers call /chat api" });
         }
-
         // process
-        const chatbot = getChatbot(req);
-        const input = getChatInput(req.body.input, provider);
+        const chatbot = ChatbotHelpers.getChatbot(req);
+        const input = ChatbotHelpers.getChatInput(req.body.input, provider);
         res.setHeader('Content-Type', 'text/event-stream');
         res.setHeader('Cache-Control', 'no-cache');
         res.setHeader('Connection', 'keep-alive');
